+++ src/pages/register/Register.tsx
@@ -62,12 +62,13 @@ export default function Register() {
 	});
 
 	const[xOffSet,setXOffSet]=useState<number>(-22);
-
+	const[showPopUp,setShowPopUp]=useState<boolean>(false);
+	const[showReg,setShowReg]=useState<boolean>(false);
 	
 
 
 	return (
-		<UserContext.Provider value={{ regInfo, setRegInfo,proPostInfo,setProPostInfo,xOffSet,setXOffSet }}>
+		<UserContext.Provider value={{ regInfo, setRegInfo,proPostInfo,setProPostInfo,xOffSet,setXOffSet,showPopUp,setShowPopUp,showReg,setShowReg }}>
 			<section className="w-[100%]">
 				<RegHeader />
 				<Outlet />
