+++ tailwind.config.js
@@ -2,7 +2,11 @@
 module.exports = {
 	content: ["./src/**/*.{js,jsx,ts,tsx}"],
 	theme: {
-		extend: {},
+		extend: {
+			colors: {
+				"black-rgba": "rgba(0, 0, 0, 0.25)",
+			},		
+		},
 		screens: {
 			"sm": "640px",
 			// => @media (min-width: 640px) { ... }
@@ -17,7 +21,7 @@ module.exports = {
 
 			"2xl": "1536px",
 			// => @media (min-width: 1536px) { ... }
-		}
+		},
 	},
 	plugins: [],
 };
